Kind of Satisfied ngl
I actually feel pretty satisfied with what this has. Of course I want to add some if and else's once I get to that part of the book. I do want to treat this as a class but it gets hard when I'm not really in a class lol. I'll get there though. I'll get to reading once I'm done here.
:+1: