O damn I didn't upload this I think yesterday
So, I've recently started to do a project about calculating your GPA. I want to someday make this for a phone or for the PC, but I'll see. I want to make it for the phone since it's waaaaay more convenient than taking out your PC. But both or just pc is fine. I also did some other code too.
For the GPA one, I've been working on the flowchart and the psuedocode. Like, I know what is supposed to happen, but making it into code is difficult. I think I have to but lists within lists. I'm not so sure though. Idk I'll see.
Ya though.
Time to play some games and then get back into this.